Merge the class and rt-class attributes
Previously an error was raised when a node had both the class and the
rt-class attribute defined. This leads to unnecessary verbose code which
taxes the reader. Splitting static class names from dynamic ones removes
this distraction, and keeps the semantics clean in the reader's mind.

On renaming .*Props to .*Attrs. I would have liked to avoid such global
renames on my first commit to somebody else's code base, but I needed to
add a constant for "className", and that seemed like a property on a
React object, as opposed to an attribute on an html element that only
later gets converted to a property. So I took the liberty to rename the
variables.

On tests. Having noticed how order-depenedent some of the code is (in
particular when it comes to dealing with rt-if/rt-repeat/rt-scope), I
decided it might be a good idea to be somewhat black-boxy in the tests,
and ensure that merging of class names works correctly in both
directions.