In rt-scope subsequent aliases should be able to reference preceding ones
Rationale: It is typical to want to alias multiple subtrees of a nested
structure in a single rt-scope statement, like so:

<div rt-scope="this.props.user as user; user.profile as profile;
user.friends as friends">

This was not possible because of the function parameter based
implementation of rt-scope. I tweaked it a bit to use var declaration
instead, while preserving the parameter-based passing for all child
scopes.