rt-scope should be evaluated in context of each rt-repeat's iteration
Rationale: similar to the change regarding the precedence of rt-if and
rt-scope, if we agree that rt-scope is mainly used to bind variables for
the use in the children, then it makes sense to have it per-iteration,
rather than as the top level over the whole rt-repeat.

This change is breaking the use case where rt-scope is used to define a
variable that is then referenced in rt-repeat. The idea is that such a
use case is rare, and the one described above is far more common.